+++ tests/testthat/test-reduce.R
@@ -24,6 +24,13 @@ test_that("direction of reduce determines how generated trees lean", {
   expect_identical(reduce(1:4, list, .dir = "backward"), list(1L, list(2L, list(3L, 4L))))
 })
 
+test_that("can shortcircuit reduction with done_box()", {
+  x <- c(TRUE, TRUE, FALSE, TRUE, TRUE)
+  out <- reduce(x, ~ if (.y) c(.x, "foo") else done_box(.x), .init = NULL)
+  expect_identical(out, c("foo", "foo"))
+})
+
+
 # accumulate --------------------------------------------------------------
 
 test_that("accumulate passes arguments to function", {
@@ -110,4 +117,10 @@ test_that("accumulate_right still works", {
 
   expect_identical(accumulate_right(0:1, c, .init = 2L), list(2:0, 2:1, 2L))
   expect_identical(accumulate_right(c(a = 0L, b = 1L), c, .init = 2L), list(b = 2:0, a = 2:1, .init = 2L))
+})
+
+test_that("can shortcircuit reduce2() with done_box()", {
+  x <- c(TRUE, TRUE, FALSE, TRUE, TRUE)
+  out <- reduce2(x, 1:5, ~ if (.y) c(.x, "foo") else done_box(.x), .init = NULL)
+  expect_identical(out, c("foo", "foo"))
 })